+++ js/form.js
@@ -1,17 +1,14 @@
 import {createBookingAd} from './data.js';
-import {setData} from './util.js';
-
-const addressField = document.querySelector('#address');
-
+import {changeState, setData} from './util.js';
 const similarCards = Array.from({length: 10}, createBookingAd);
+const addressField = document.querySelector('#address');
 const cardTemplate = document.querySelector('#card')
   .content
   .querySelector('.popup');
-// const mapCanvas = document.querySelector('#map-canvas');
-
+const sliderElement = document.querySelector('#slider');
+const valueElement = document.querySelector('#price');
 const spawnCard = function (data) {
   const cardElement = cardTemplate.cloneNode(true);
-
   const popupTitle = cardElement.querySelector('.popup__title');
   const popupTextAdress = cardElement.querySelector('.popup__text--address');
   const popupTextPrice = cardElement.querySelector('.popup__text--price');
@@ -21,7 +18,6 @@ const spawnCard = function (data) {
   const popupDescription = cardElement.querySelector('.popup__description');
   const popupPhoto = cardElement.querySelector('.popup__photo');
   const popupAvatar = cardElement.querySelector('.popup__avatar');
-
   const popupFeatures = cardElement.querySelector('.popup__features')
     .querySelectorAll('.popup__feature');
   const modifiers = data.offer.features.map((feature) => `popup__feature--${  feature}`);
@@ -42,20 +38,19 @@ const spawnCard = function (data) {
   setData(popupTextTime,[data.offer.checkin, data.offer.checkout], 'textContent', `Заезд после ${data.offer.checkin}, выезд до ${data.offer.checkout}`);
   return cardElement;
 };
-similarCards.forEach((element) => spawnCard(element));
 // LEAFLET
 const map = L.map('map-canvas').setView(
   {
     lat: 35.652832,
     lng: 139.839478,
   }, 10);
-
 L.tileLayer(
   'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
   {
     attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
   },
 ).addTo(map);
+map.on('load', changeState(1));
 const mainPinIcon = L.icon({
   iconUrl: './img/main-pin.svg',
   iconSize: [52, 52],
@@ -80,7 +75,6 @@ marker.on('moveend', (evt) => {
   const latLng = evt.target.getLatLng();
   addressField.value = `${latLng.lat.toFixed(5)  } ${  latLng.lng.toFixed(5)}`;
 });
-
 marker.addTo(map);
 
 const createAdds = (element) => {
@@ -100,4 +94,18 @@ const createAdds = (element) => {
 
 similarCards.forEach((element) => {
   createAdds(element);
+});
+// noUiSlider
+noUiSlider.create(sliderElement, {
+  range: {
+    min: 0,
+    max: 100000,
+  },
+  start: 0,
+  step: 100,
+  connect: 'lower',
+});
+
+sliderElement.noUiSlider.on('update', () => {
+  valueElement.value = sliderElement.noUiSlider.get();
 });
