--- js/pagestate.js
@@ -1,18 +1,0 @@
-let addressField = document.querySelector('#address');
-
-
-const changeState = function (state) {
-  const adForm = document.querySelector('.ad-form');
-  const mapFilters = document.querySelector('.map__filters');
-  if (state === 0) {
-    adForm.classList.add('ad-form--disabled');
-    adForm.setAttribute('disabled', 'disabled');
-    mapFilters.classList.add('map__filters--disabled');
-    mapFilters.setAttribute('disabled', 'disabled');
-  }  else if (state === 1) {
-    adForm.classList.remove('ad-form--disabled');
-    adForm.removeAttribute('disabled');
-    mapFilters.classList.remove('map__filters--disabled');
-    mapFilters.removeAttribute('disabled');
-  }};
-changeState(1);
